utils/factories.py
import datetime
import factory
from django.contrib.auth.hashers import make_password
from django.utils import timezone

from appointment.models import CardDoctor, Appointment
from user.models import Profile

