requirements.txt
celery==5.1.2
Django==3.2.7
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
Pillow==8.3.2
psycopg2-binary==2.9.1
python-dotenv==0.19.0
factory_boy==3.2.0
